Add test for verified boards
This commit adds a test at t/verify.t which uses the board text files
in the verified/ subdirectory, which are machine-readable outputs of
blokus.pl that were checked by a human for correctness, to make sure
that the output of blokus.pl on the same image is still the same board.

This test should be run whenever the G'MIC filters or cluster centers
in the MPI::Blokus::* modules are changed, to ensure that such changes
do not break what worked before. The test suite can be run via `prove`:

    $ prove t/