Clone cached color centers
The color centers for the cluster_colors in MPI::Blokus::Colors are computed
from samples when they are first needed and then stored in `state` variables.
However, those state variables are arrayrefs and they are manipulated by the
clustering algorithm directly. This leaves the centers of a previous cluster
run as the new start centers for the next run which is wrong because another
photo may have different lighting, a different base color, etc. The fix is
to simply `clone` the state variable.